requirements.txt
telethon
pandas
python-docx
lxml
openpyxl
xlsxwriter
pytz
cryptography
python-dotenv